# Replace reverse.each with reverse_each